Fix bug with AutoDJ updating LastPlayed erroneously
The old random-bys used the LastPlayed/LastSkipped values in CoreTracks
to not repeat a track that had already been shuffled.  This works well
for playback shuffling, but the AutoDJ was using it for its
fill-shuffling as well.  Fix that by adding a CoreShuffles table that
the PlayQueue and others can use to get a random track from a source,
without repeats, and without messing with the LastPlayedStamp.  This
fixes BGO #594701