[GStreamer] Don't track volume in preferences
Added a new bp_supports_stream_volume API that returns whether or not
the playbin element supports GST_TYPE_STREAM_VOLUME. If stream volumes
are supported on the pipeline (which, for example, map to the system
mixer), then don't load/save the volume from Banshee's preferences and
instead only load/set volume on the pipeline itself so the application
volume state is always in sync with the pipeline.

This *should* fix bgo#613725, bgo#612362, and bgo#613726, though there
may be some sort of inconsistency on some distributions as this fix will
only apply for GStreamer 0.10.25 or newer (when the
GST_TYPE_STREAM_VOLUME interface was introduced). If the case is that
stream volume was supported internally on playbin2 before 0.10.25 adds
this interface, there's no real way for us to know whether to trust the
pipeline (system/mixer) volume or our internal software volume state.