[PlayQueueSource] Use localized name as special name
Use the translation of "Play Queue" as the special name to recognize the
playlist in the CorePlaylists table, so that it's the same value as the
playlist name. If we don't do that, confusion arises and the content of
the play queue is lost between restarts when using a non english locale.
Fixes bgo#621480.