[libbanshee] Fix up some volume handling issues
Check the audiosink recursively for a 'volume' property.  If it exists,
then we assume the sink will save/restore its own volume - so expose
that information to the managed PlayerEngine so it can decide whether to
restore the last volume from gconf.  Also, listen to the
deep-notify::volume signal like RB does.