AudioCd: Fix metadata fetching with multiple results (bgo#587264)
When looking for CD metadata on MusicBrainz, we were looking for a
perfect match: only one release with a score of 100. If the search
returned more than one result with a score of 100, we would reject all
results.
MusicBrainz recently started to return multiple releases much more
often, causing our metadata fetching to fail quite a lot.

We now use the first result if it has a score of 100, ignoring any other
results.

This also fixes a NRE when no results were found.